refactored decorate to Engine
